+++ src/backend/marsha/core/tests/api/video/test_create.py
@@ -187,8 +187,51 @@ class VideoCreateAPITest(TestCase):
             HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
         )
 
-        self.assertEqual(models.Video.objects.count(), 0)
-        self.assertEqual(response.status_code, 403)
+        self.assertEqual(models.Video.objects.count(), 1)
+        self.assertEqual(response.status_code, 201)
+        self.assertEqual(
+            response.json(),
+            {
+                "active_shared_live_media": None,
+                "active_shared_live_media_page": None,
+                "active_stamp": None,
+                "allow_recording": True,
+                "description": "",
+                "estimated_duration": None,
+                "has_chat": True,
+                "has_live_media": True,
+                "has_transcript": False,
+                "id": str(models.Video.objects.get().id),
+                "is_public": False,
+                "is_ready_to_show": False,
+                "is_recording": False,
+                "is_scheduled": False,
+                "join_mode": "approval",
+                "live_info": {},
+                "live_state": None,
+                "live_type": None,
+                "participants_asking_to_join": [],
+                "participants_in_discussion": [],
+                "playlist": {
+                    "id": str(playlist.id),
+                    "lti_id": playlist.lti_id,
+                    "title": playlist.title,
+                },
+                "recording_time": 0,
+                "shared_live_medias": [],
+                "should_use_subtitle_as_transcript": False,
+                "show_download": True,
+                "starting_at": None,
+                "thumbnail": None,
+                "timed_text_tracks": [],
+                "title": "Some video",
+                "upload_state": "pending",
+                "urls": None,
+                "xmpp": None,
+                "tags": [],
+                "license": None,
+            },
+        )
 
     def test_api_video_create_by_organization_admin(self):
         """