+++ src/backend/marsha/core/tests/api/video/test_start_live.py
@@ -8,7 +8,7 @@ from django.test import TestCase, override_settings
 
 from boto3.exceptions import Boto3Error
 
-from marsha.core import api, factories
+from marsha.core import api, factories, models
 from marsha.core.api import timezone
 from marsha.core.defaults import (
     HARVESTED,
@@ -23,7 +23,10 @@ from marsha.core.defaults import (
     STOPPED,
 )
 from marsha.core.factories import LiveSessionFactory
-from marsha.core.simple_jwt.factories import InstructorOrAdminLtiTokenFactory
+from marsha.core.simple_jwt.factories import (
+    InstructorOrAdminLtiTokenFactory,
+    UserAccessTokenFactory,
+)
 from marsha.core.utils.time_utils import to_timestamp
 
 
@@ -31,6 +34,156 @@ class VideoStartLiveAPITest(TestCase):
     """Test the "start live" API of the video object."""
 
     maxDiff = None
+
+    @classmethod
+    def setUpTestData(cls):
+        super().setUpTestData()
+
+        cls.some_organization = factories.OrganizationFactory()
+        cls.some_video = factories.WebinarVideoFactory(
+            playlist__organization=cls.some_organization,
+        )
+
+    def assert_user_cannot_start_live(self, user, video):
+        """Assert the user cannot start the live."""
+
+        jwt_token = UserAccessTokenFactory(user=user)
+        response = self.client.post(
+            f"/api/videos/{video.pk}/start-live/",
+            HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
+        )
+
+        self.assertEqual(response.status_code, 403)
+
+    @override_settings(LIVE_CHAT_ENABLED=True)
+    def assert_user_can_start_live(self, user, video):
+        """Assert the user can start the live."""
+        self.assertNotEqual(video.live_state, STARTING)
+
+        with mock.patch.object(api.video, "start_live_channel"), mock.patch.object(
+            api.video, "create_live_stream"
+        ) as mock_create_live_stream, mock.patch.object(
+            api.video, "wait_medialive_channel_is_created"
+        ) as mock_wait_medialive_channel_is_created, mock.patch(
+            "marsha.core.serializers.xmpp_utils.generate_jwt"
+        ) as mock_jwt_encode, mock.patch.object(
+            api.video, "create_room"
+        ) as mock_create_room, mock.patch(
+            "marsha.websocket.utils.channel_layers_utils.dispatch_video_to_groups"
+        ) as mock_dispatch_video_to_groups:
+            mock_jwt_encode.return_value = "xmpp_jwt"
+            mock_create_live_stream.return_value = {
+                "medialive": {
+                    "input": {
+                        "id": "medialive_input_1",
+                        "endpoints": [
+                            "https://live_endpoint1",
+                            "https://live_endpoint2",
+                        ],
+                    },
+                    "channel": {"id": "medialive_channel_1"},
+                },
+                "mediapackage": {
+                    "id": "mediapackage_channel_1",
+                    "endpoints": {
+                        "hls": {
+                            "id": "endpoint1",
+                            "url": "https://channel_endpoint1/live.m3u8",
+                        },
+                    },
+                },
+            }
+            jwt_token = UserAccessTokenFactory(user=user)
+            response = self.client.post(
+                f"/api/videos/{video.id}/start-live/",
+                HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
+            )
+            mock_create_room.assert_called_with(video.id)
+            mock_wait_medialive_channel_is_created.assert_called_with(
+                "medialive_channel_1"
+            )
+            mock_dispatch_video_to_groups.assert_has_calls(
+                [mock.call(video), mock.call(video)]
+            )
+
+        self.assertEqual(response.status_code, 200)
+        content = json.loads(response.content)
+
+        self.assertEqual(content["live_state"], STARTING)
+
+    def test_start_live_by_anonymous_user(self):
+        """Anonymous users cannot start a live."""
+        response = self.client.post(f"/api/videos/{self.some_video.pk}/start-live/")
+
+        self.assertEqual(response.status_code, 401)
+
+    def test_start_live_by_random_user(self):
+        """Authenticated user without access cannot start a live."""
+        user = factories.UserFactory()
+
+        self.assert_user_cannot_start_live(user, self.some_video)
+
+    def test_start_live_by_organization_student(self):
+        """Organization students cannot start a live."""
+        organization_access = factories.OrganizationAccessFactory(
+            organization=self.some_organization,
+            role=models.STUDENT,
+        )
+
+        self.assert_user_cannot_start_live(organization_access.user, self.some_video)
+
+    def test_start_live_by_organization_instructor(self):
+        """Organization instructors cannot start a live."""
+        organization_access = factories.OrganizationAccessFactory(
+            organization=self.some_organization,
+            role=models.INSTRUCTOR,
+        )
+
+        self.assert_user_cannot_start_live(organization_access.user, self.some_video)
+
+    def test_start_live_by_organization_administrator(self):
+        """Organization administrators can start a live."""
+        organization_access = factories.OrganizationAccessFactory(
+            organization=self.some_organization,
+            role=models.ADMINISTRATOR,
+        )
+
+        self.assert_user_can_start_live(organization_access.user, self.some_video)
+
+    def test_start_live_by_consumer_site_any_role(self):
+        """Consumer site roles cannot start a live."""
+        consumer_site_access = factories.ConsumerSiteAccessFactory(
+            consumer_site=self.some_video.playlist.consumer_site,
+        )
+
+        self.assert_user_cannot_start_live(consumer_site_access.user, self.some_video)
+
+    def test_start_live_by_playlist_student(self):
+        """Playlist student cannot start a live."""
+        playlist_access = factories.PlaylistAccessFactory(
+            playlist=self.some_video.playlist,
+            role=models.STUDENT,
+        )
+
+        self.assert_user_cannot_start_live(playlist_access.user, self.some_video)
+
+    def test_start_live_by_playlist_instructor(self):
+        """Playlist instructor cannot start a live."""
+        playlist_access = factories.PlaylistAccessFactory(
+            playlist=self.some_video.playlist,
+            role=models.INSTRUCTOR,
+        )
+
+        self.assert_user_can_start_live(playlist_access.user, self.some_video)
+
+    def test_start_live_by_playlist_admin(self):
+        """Playlist administrator can start a live."""
+        playlist_access = factories.PlaylistAccessFactory(
+            playlist=self.some_video.playlist,
+            role=models.ADMINISTRATOR,
+        )
+
+        self.assert_user_can_start_live(playlist_access.user, self.some_video)
 
     @override_settings(LIVE_CHAT_ENABLED=True)
     @override_settings(XMPP_BOSH_URL="https://xmpp-server.com/http-bind")