+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add a check on deposited file size when uploading content
 - helpers frontend api error handling
 - Add accepted formats in the subtitles uploaders helptext
+- Standalone website:
+  - Add "video" management API endpoints permissions
 
 ### Changed